Renamed `processed_block_trace.rs` --> `processed_block_trace_mpt.rs`
- Prep for `mpt`/`smt` logic feature gating.
